+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 * Intial release
 * Supervisor device endpoint node
+
+## 0.0.2
+* changed nodes color and icon
+* fixed a bug where Supervisor device endpoint node would return only `state`
+* Supervisor ping endpoint node